+++ CHANGELOG.md
@@ -6,8 +6,8 @@ Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` w
 
 - Major: Added support for 7tv.app emotes (#1256, #1260, #1261, #1317)
 - Minor: Added `can_execute_with_whisper`, `sub_only`, `mod_only` & `run_through_banphrases` fields to create command page. (#1310)
-- Minor: Added support for the `extendsub` msg-id to the `subalert` module.
-- Minor: Added support for the `giftpaidupgrade` msg-id to the `subalert` module. (#1306, #1315)
+- Minor: Added support for the `extendsub` msg-id to the `subalert` module. (#1311, #1318)
+- Minor: Added support for the `giftpaidupgrade` msg-id to the `subalert` module. (#1306, #1315, #1318)
 - Minor: Migrate from Kraken to Helix for global emote fetching. (#1289)
 - Minor: Migrate from `twitchemotes.com` to Helix for channel emote fetching. (#1290)
 - Minor: Added `mod_only` & `run_through_banphrases` fields to edit command page. (#1293)